Presentation on Research
This presentation was given at the PNW Annual meeting of the Mathematical Association of America as well as Founder's Day at the University of Portland. Overall, the presentation summarizes the key findings of our research over the last year and a half as well as explains the statistical tools used to arrive at key conclusions.